Move spell implementation into separate file
