+++ src/fheroes2/ai/normal/ai_normal_battle.cpp
@@ -78,124 +78,6 @@ namespace AI
         return currentUnit.isFlying();
     }
 
-    SpellSeletion BattlePlanner::selectBestSpell( Battle::Arena & arena, bool retreating )
-    {
-        // 1. For damage spells - maximum amount of enemy threat lost
-        // 2. For buffs - friendly unit strength gained
-        // 3. For debuffs - enemy unit threat lost
-        // 4. For dispell, resurrect and cure - amount of unit strength recovered
-        // 5. For antimagic - based on enemy hero spellcasting abilities multiplied by friendly unit strength
-
-        // 6. Cast best spell with highest heuristic on target pointer saved
-        SpellSeletion bestSpell;
-
-        // Commander must be set before calling this function! Check both debug/release version
-        assert( _commander != nullptr );
-        if ( _commander == nullptr ) {
-            return bestSpell;
-        }
-
-        const std::vector<Spell> allSpells = _commander->GetSpells();
-        const Units friendly( arena.GetForce( _myColor ), true );
-        const Units enemies( arena.GetForce( _myColor, true ), true );
-
-        double bestHeuristic = 0;
-
-        const int spellPower = _commander->GetPower();
-        for ( const Spell & spell : allSpells ) {
-            if ( !_commander->HaveSpellPoints( spell ) || !spell.isCombat() || ( !spell.isDamage() && retreating ) )
-                continue;
-
-            if ( spell.isDamage() ) {
-                const SpellcastOutcome & outcome = spellDamageValue( spell, arena, friendly, enemies, retreating );
-                if ( outcome.value > bestHeuristic ) {
-                    bestHeuristic = outcome.value;
-                    bestSpell.spellID = spell.GetID();
-                    bestSpell.cell = outcome.cell;
-                }
-            }
-        }
-        return bestSpell;
-    }
-
-    SpellcastOutcome BattlePlanner::spellDamageValue( const Spell & spell, Arena & arena, const Units & friendly, const Units & enemies, bool retreating ) 
-    {
-        SpellcastOutcome bestOutcome;
-        if ( !spell.isDamage() )
-            return bestOutcome;
-
-        const int spellPower = _commander->GetPower();
-        const uint32_t totalDamage = spell.Damage() * spellPower;
-
-        auto damageHeuristic = [&totalDamage, &spell, &spellPower, &retreating]( const Unit * unit ) {
-            const uint32_t damage = totalDamage * ( 100 - unit->GetMagicResist( spell, spellPower ) ) / 100;
-            // If we're retreating we don't care about partial damage, only actual units killed
-            if ( retreating )
-                return unit->GetMonsterStrength() * unit->HowManyWillKilled( damage );
-            // Otherwise calculate amount of strength lost (% of unit times total strength)
-            return std::min( static_cast<double>( damage ) / unit->GetHitPoints(), 1.0 ) * unit->GetStrength();
-        };
-
-        if ( spell.isSingleTarget() ) {
-            for ( const Unit * enemy : enemies ) {
-                const double spellHeuristic = damageHeuristic( enemy );
-
-                if ( spellHeuristic > bestOutcome.value ) {
-                    bestOutcome.value = spellHeuristic;
-                    bestOutcome.cell = enemy->GetHeadIndex();
-                }
-            }
-        }
-        else if ( spell.isApplyWithoutFocusObject() ) {
-            double spellHeuristic = 0;
-            for ( const Unit * enemy : enemies ) {
-                spellHeuristic += damageHeuristic( enemy );
-            }
-            for ( const Unit * unit : friendly ) {
-                spellHeuristic -= damageHeuristic( unit );
-            }
-
-            if ( spellHeuristic > bestOutcome.value ) {
-                bestOutcome.value = spellHeuristic;
-            }
-        }
-        else {
-            // Area of effect spells like Fireball
-            auto areaOfEffectCheck = [&damageHeuristic, &bestOutcome]( const TargetsInfo & targets, const int32_t index, int myColor ) {
-                double spellHeuristic = 0;
-                for ( const TargetInfo & target : targets ) {
-                    if ( target.defender->GetCurrentColor() == myColor ) {
-                        spellHeuristic -= damageHeuristic( target.defender );
-                    }
-                    else {
-                        spellHeuristic += damageHeuristic( target.defender );
-                    }
-                }
-
-                if ( spellHeuristic > bestOutcome.value ) {
-                    bestOutcome.value = spellHeuristic;
-                    bestOutcome.cell = index;
-                }
-            };
-
-            if ( spell.GetID() == Spell::CHAINLIGHTNING ) {
-                for ( const Unit * enemy : enemies ) {
-                    const int32_t index = enemy->GetHeadIndex();
-                    areaOfEffectCheck( arena.GetTargetsForSpells( _commander, spell, index, false ), index, _myColor );
-                }
-            }
-            else {
-                const Board & board = *Arena::GetBoard();
-                for ( const Cell & cell : board ) {
-                    const int32_t index = cell.GetIndex();
-                    areaOfEffectCheck( arena.GetTargetsForSpells( _commander, spell, index, false ), index, _myColor );
-                }
-            }
-        }
-
-        return bestOutcome;
-    }
-
     Actions BattlePlanner::planUnitTurn( Arena & arena, const Unit & currentUnit )
     {
         if ( currentUnit.Modes( SP_BERSERKER ) != 0 ) {
